Added patch for fixing Member role
https://review.rdoproject.org/r/#/c/20579/ updates the tempestconf
version to 2.2.0 to queens, but this release brings regression
in the queens tempestconf generation leads to block queens
promotion and upstream tripleo ci job.

https://review.opendev.org/#/c/660748/ fixes the same but it is
in gates and got blocked on other gate issue. Once the patch
merged, also it will come in a new tag.

Currently cherry-picking this patch with the spec file unblocks
the gates and fixes the issue.

https://bugs.launchpad.net/tripleo/+bug/1828859

Change-Id: Ia2296e3f94880442e1800ac9261f762a8b0fe91c
Signed-off-by: Chandan Kumar (raukadah) <chkumar246@gmail.com>